Add: 0-read_textfile.c

+++ 0x15-file_io/0-read_textfile.c
@@ -5,7 +5,8 @@
 #include <stdlib.h>
 
 /**
- * read_textfile - his function reads a text file and prints it to the POSIX standard output
+ * read_textfile - his function reads a text file and prints it to the POSIX
+ *  standard output
  * @filename: Name of the file to print
  * @letters: Number of letters it should read and print
  *
@@ -16,7 +17,7 @@ ssize_t read_textfile(const char *filename, size_t letters)
 	int fd;
 	int res;
 	char *buffer = malloc(sizeof(char) * letters);
-	ssize_t noChars;
+	int noChars;
 
 	if (!filename)
 		return (0);
